Include whole directories in final package
The current approach of only including files in the final RPM leads to
leaving stale directories behind after uninstallation. That's because
the package manager assumes that the directories are not part of the
package, only certain contents, and it leaves them as is.

Simplify our spec file and always include directories, to make the
package manager remove them when we uninstall the package.