Download PySide6 wheels from Qt snapshots
The wheels for the 6.6.3.1 release are not available from the official
Qt downloads page [1]. However, they are available from the snapshots
page [2], and are most probably uploaded by Qt's CI runners.

We have verified locally that the hashes that PyPI provides for 6.6.3.1
match the hashes of the PySide6 wheels, as downloaded from the snapshots
page. For this reason, we feel confident that we can switch for this
release [3] the download URL from the official downloads page to the
snapshots one.

[1]: https://download.qt.io/official_releases/QtForPython/pyside6/
[2]: https://download.qt.io/snapshots/ci/pyside/
[3]: This release will most probably be the last one, see issue #5